requirements.txt
-e git+ssh://git@github.com/aptivate/django.js.git@ca328a94b00023bd64f4fc1c908675edaaf2ac19#egg=django.js

Django==1.11.15
MySQL-python==1.2.5
cssmin==0.2.0
django-assets==0.12
django-bootstrap3==11.0.0
django-braces==1.13.0
django-filter==1.1.0
django-floppyforms==1.7.0
django-form-utils==1.0.3
django-spreadsheetresponsemixin==0.1.10
django-tables2==1.21.2
django-widget-tweaks==1.4.2
djangorestframework-bulk==0.2.1
djangorestframework==3.8.2
factory_boy==2.11.1
jsmin==2.2.2
jsonfield==2.0.2
linecache2==1.0.0
mock==2.0.0
openpyxl==2.5.6
pillow==5.2.0
py==1.6.0
pytest-django==3.4.2
pytest-pythonpath==0.7.3
python-dateutil==2.7.3
pytz==2018.5
isort==4.3.4
pylava==0.2.1
